Initial commit to move CredentialsProvider under athena-federation-sdk
