+++ athena-federation-sdk/src/test/java/com/amazonaws/athena/connector/credentials/JdbcCredentialProviderTest.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.amazonaws.athena.connectors.jdbc.connection;
+package com.amazonaws.athena.connector.credentials;
 
 import org.junit.Assert;
 import org.junit.Test;
